BIG types and mocks refactor and division landing page
