+++ __tests__/__mocks__/data/mockCollections.tsx
@@ -1,0 +1,60 @@
+import CollectionDataType from "src/types/CollectionDataType";
+
+export const mockCollections: CollectionDataType[] = [
+  {
+    uuid: "60932400-20f2-0138-8583-05c43d448773",
+    title: "Posada Collection",
+    url: "https://digitalcollections.nypl.org/collections/posada-collection#/?tab=navigation",
+    image_id: "58270299",
+    numItems: 34,
+  },
+  {
+    uuid: "724303e0-c6bb-012f-afbd-58d385a7bc34",
+    title: "MAVO",
+    url: "https://digitalcollections.nypl.org/collections/mavo#/?tab=navigation",
+    image_id: "1408153",
+    numItems: 35,
+  },
+  {
+    uuid: "6b6532b0-5df7-013b-36f8-0242ac110002",
+    title: "Austin Hansen photograph collection",
+    url: "https://digitalcollections.nypl.org/collections/austin-hansen-photograph-collection#/?tab=navigation",
+    image_id: "58300996",
+    numItems: 65,
+  },
+  {
+    uuid: "5b996640-c31c-0139-0bac-0242ac110004",
+    title: "Arthur Alfonso Schomburg papers",
+    url: "https://digitalcollections.nypl.org/collections/arthur-alfonso-schomburg-papers#/?tab=navigation",
+    image_id: "58591658",
+    numItems: 55,
+  },
+  {
+    uuid: "d3802d10-f49a-0139-3bff-0242ac110002",
+    title: "Friedman-Abeles photographs",
+    url: "https://digitalcollections.nypl.org/collections/friedman-abeles-photographs#/?tab=navigation",
+    image_id: "58498722",
+    numItems: 35,
+  },
+  {
+    uuid: "e5462600-c5d9-012f-a6a3-58d385a7bc34",
+    title: "Farm Security Administration Photographs",
+    url: "https://digitalcollections.nypl.org/collections/farm-security-administration-photographs#/?tab=navigation",
+    image_id: "1952272",
+    numItems: 36,
+  },
+  {
+    uuid: "812e5770-c60c-012f-7167-58d385a7bc34",
+    title: "Changing New York",
+    url: "https://digitalcollections.nypl.org/collections/changing-new-york#/?tab=navigation",
+    image_id: "58447105",
+    numItems: 37,
+  },
+  {
+    uuid: "f7ffc990-c5ae-012f-eb75-58d385a7bc34",
+    title: "Diana Davies photographs",
+    url: "https://digitalcollections.nypl.org/collections/diana-davies-photographs#/?tab=navigation",
+    image_id: "1582202",
+    numItems: 38,
+  },
+];
