+++ src/components/hero/campaignHero.tsx
@@ -5,7 +5,7 @@ import CampaignHeroHeading from "./campaignHeroHeading";
 import CampaignHeroLoading from "./campaignHeroLoading";
 import defaultFeaturedItem from "../../data/defaultFeaturedItemData";
 import appConfig from "../../../appConfig";
-import { FeaturedItemDataType } from "../../types/FeaturedItemData";
+import { FeaturedItemDataType } from "../../types/FeaturedItemDataType";
 import React from "react";
 import { ENV_KEY } from "src/types/EnvironmentType";
