+++ src/components/swimlanes/swimlanes.test.tsx
@@ -1,14 +1,20 @@
 import { fireEvent, render, screen, within } from "@testing-library/react";
 import SwimLanes from "./swimLanes";
-import { props } from "../../../__tests__/data/homepageProps";
+import { mockSwimLanes } from "__tests__/__mocks__/data/mockSwimLanes";
 
 describe("Swim Lanes component renders with expected props", () => {
   it("renders the first row", () => {
-    render(<SwimLanes {...props} />);
+    render(
+      <SwimLanes
+        numColumns={4}
+        lanesWithNumItems={mockSwimLanes.lanesWithNumItems}
+      />
+    );
     const firstrow = screen.getByTestId("test-collections-1");
     expect(within(firstrow).getByText("Posada Collection")).toBeInTheDocument();
     expect(within(firstrow).getByText("34 items")).toBeInTheDocument();
     expect(within(firstrow).findAllByRole("img")).toBeVisible;
+    console.log(within(firstrow).getAllByRole("img"));
     expect(within(firstrow).getByText("MAVO")).toBeInTheDocument();
     expect(
       within(firstrow).getByText("Austin Hansen photograph collection")
@@ -19,7 +25,12 @@ describe("Swim Lanes component renders with expected props", () => {
   });
 
   it("renders the second row", () => {
-    render(<SwimLanes {...props} />);
+    render(
+      <SwimLanes
+        numColumns={4}
+        lanesWithNumItems={mockSwimLanes.lanesWithNumItems}
+      />
+    );
     const secondrow = screen.getByTestId("test-collections-2");
     expect(
       within(secondrow).getByText("Friedman-Abeles photographs")
@@ -38,18 +49,28 @@ describe("Swim Lanes component renders with expected props", () => {
   });
 
   it("renders tooltips on >1024px width", async () => {
-    render(<SwimLanes {...props} />);
+    render(
+      <SwimLanes
+        numColumns={4}
+        lanesWithNumItems={mockSwimLanes.lanesWithNumItems}
+      />
+    );
     window.innerWidth = 1050;
     fireEvent(window, new Event("resize"));
-    fireEvent.pointerOver(screen.getByText("MAVO"));
+    fireEvent.pointerOver(screen.getAllByText("MAVO")[0]);
     const tooltip = await screen.findByRole("tooltip");
     expect(tooltip).toBeInTheDocument();
   });
   it("does not render tooltips <1024px width", async () => {
-    render(<SwimLanes {...props} />);
+    render(
+      <SwimLanes
+        numColumns={4}
+        lanesWithNumItems={mockSwimLanes.lanesWithNumItems}
+      />
+    );
     window.innerWidth = 1000;
     fireEvent(window, new Event("resize"));
-    fireEvent.pointerOver(screen.getByText("MAVO"));
+    fireEvent.pointerOver(screen.getAllByText("MAVO")[0]);
     expect(screen.queryByRole("tooltip")).not.toBeInTheDocument();
   });
 });
