+++ src/types/CollectionDataType.ts
@@ -1,0 +1,10 @@
+export interface CollectionDataType {
+  uuid: string;
+  title: string;
+  url: string;
+  image_id: string;
+  numItems: number;
+  // TODO: containsOnSiteMaterials: boolean; // to be added in phase 2
+}
+
+export default CollectionDataType;
