+++ src/types/ExploreFurtherDataType.ts
