+++ src/types/FeaturedContentDataType.ts
