+++ src/types/props/SwimLanesProps.ts
@@ -1,0 +1,11 @@
+import CollectionDataType from "../CollectionDataType";
+
+export interface SwimLanesProps {
+  numColumns: number;
+  lanesWithNumItems: {
+    slug: string;
+    title: string;
+    rank: number;
+    collections: CollectionDataType[];
+  }[];
+}
